+++ packages/author-head/author-head.js
@@ -15,7 +15,7 @@ import AuthorHeadContainer from "./author-head-container";
 
 const styles = StyleSheet.create({
   twitter: {
-    paddingTop: 16,
+    paddingTop: 8,
     paddingBottom: 8,
     flexDirection: "row",
     ...Platform.select({