+++ packages/author-profile/author-profile-item.js
@@ -8,8 +8,8 @@ import ArticleSummary from "@times-components/article-summary";
 
 const styles = StyleSheet.create({
   container: {
-    paddingBottom: 10,
-    paddingTop: 10
+    paddingBottom: 15,
+    paddingTop: 15
   }
 });