+++ packages/gestures/__tests__/android/gestures.android.test.js
@@ -1,0 +1,14 @@
+import tests from "../gestures.native.test";
+
+jest.mock("react-native", () => {
+  const reactNative = require.requireActual("react-native");
+  reactNative.Platform.OS = "android";
+  jest
+    .spyOn(reactNative.Platform, "select")
+    .mockImplementation(obj => obj.android || obj.default);
+  return reactNative;
+});
+
+describe("Gesture tests", () => {
+  tests();
+});