+++ packages/image/__tests__/web/jest.config.js
@@ -2,5 +2,6 @@ const jestConfigurator = require("@times-components/jest-configurator").default;
 
 module.exports = jestConfigurator("web", __dirname, [
   "add-missing-protocol.js",
-  "style-native.js"
+  "style-native.js",
+  "gestures.js"
 ]);