+++ packages/pagination/__tests__/pagination-icons.native.test.js
@@ -1,3 +1,0 @@
-import tests from "./test-icons-helper";
-
-describe("PaginationIcons Native", tests);