+++ packages/article-flag/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.10.4"></a>
+## [0.10.4](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.10.3...@times-components/article-flag@0.10.4) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
 <a name="0.10.3"></a>
 ## [0.10.3](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.10.2...@times-components/article-flag@0.10.3) (2018-01-19)