+++ packages/brightcove-video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.10.9"></a>
+## [1.10.9](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@1.10.8...@times-components/brightcove-video@1.10.9) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
 <a name="1.10.8"></a>
 ## [1.10.8](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@1.10.7...@times-components/brightcove-video@1.10.8) (2018-01-19)