+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.16"></a>
+## [0.7.16](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.7.15...@times-components/error-view@0.7.16) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
 <a name="0.7.15"></a>
 ## [0.7.15](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.7.14...@times-components/error-view@0.7.15) (2018-01-19)