+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.4"></a>
+## [0.1.4](https://github.com/newsuk/times-components/compare/@times-components/icons@0.1.3...@times-components/icons@0.1.4) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/icons
+
 <a name="0.1.3"></a>
 ## [0.1.3](https://github.com/newsuk/times-components/compare/@times-components/icons@0.1.2...@times-components/icons@0.1.3) (2018-01-19)